Fix warnings in demo project
1. Fix warnings in ‘SSLunarDate’
2. Update Readme
3. Fix warnings while orientation is changing in iPad
4. Bump version 1.6.3

+++ Example/SSLunarDate/SSHolidayWest.m
@@ -285,7 +285,7 @@
     // if that day was monday, it equal to zero.
     // if that day was tue, it equal to 1, minor 1 =
     NSDateComponents *ccc = [NSDateComponents new];
-    int reminder = cc.weekday - 2;
+    int reminder = (int)cc.weekday - 2;
     if (reminder >= 0)
         ccc.day  = - (cc.weekday - 2);
     else
